[enrich][git] Add effort pair programming metrics
pair_programming_commit with value 1 or 0.5, even other values could be allowed if there are more authors for that commit. This way, a commit in PP should count 0.5 for each author.
pair_programming_lines_*, where * can be added, removed or changed. The same as above, but for number of lines instead of commits.